added tests for registration constants
